+++ pymatviz/utils/__init__.py
@@ -2,6 +2,10 @@ from __future__ import annotations
 
 import os
 
+
 PKG_DIR = os.path.dirname(__file__)
 ROOT = os.path.dirname(PKG_DIR)
 
+
+class ExperimentalWarning(Warning):
+    """Warning for experimental features."""
