+++ pymatviz/utils/testing.py
@@ -2,4 +2,5 @@ from __future__ import annotations
 
 from pymatviz.utils import ROOT
 
+
 TEST_FILES: str = f"{ROOT}/tests/files"
